Build a docker image to use for running bdcs-api-server
Changes the build image name to bdcs-api-build-img (consistent with the
naming for bdcs-build-img), and uses that as the basis for an image
running bdcs-api-server with default paths for mddb (/mddb), store
(/mddb/cs.repo/), and recipes (/recipes).

Requires a mddb and content store to be mounted by the run command.

sudo docker run -it --rm --security-opt="label=disable" \
-v bdcs-mddb-volume:/mddb/ -v $HOME/tmp/recipes:/recipes/ welder/bdcs-api-img